[LockedData] About to write the copy constructors
About to write the copy constructors with RAII decorated constructor
delegation.  This should ideally be done with tags for maximum readability.

RAII based constructor decoration is a idiom that is used to execute code in
constructors but yet perform some action before and after the construction.
Since constructors are the only functions that can be called while leaving
objects in non-constructed states in a constructor some task can be performed
before the initialization of members in constructors by delegating
construction to an internal constructor