[googletest] Moved gtest to a separate folder
Moved the third party dependency to a seperate folder to eliminate dependence
on the position of the folder in the directory tree of the parent folder.  As
should be the case with buck modules, especially third party ones?

Only major change in the BUCK file for googletest was the it now has an empty
header_namespace argument implying that there is no positional dependence.