[Traits] Added Unique trait
sharp::Unique lets you determine whether a type list has any repeating
element.  This can be used in cases for example std::variant, it makes no
sense in std::variant to have duplicate types, i.e. std::variant<int, int>
does not make much sense, and this trait can be used as the backend of a
template that ensures this as well as an individual static_assert for better
documentation of type list invariants